Make use of UE5's double-precision

+++ Source/CesiumRuntime/Private/CesiumSunSky.cpp
@@ -470,7 +470,7 @@ void ACesiumSunSky::UpdateAtmosphereRadius() {
 }
 
 void ACesiumSunSky::GetHMSFromSolarTime(
-    float InSolarTime,
+    double InSolarTime,
     int32& Hour,
     int32& Minute,
     int32& Second) {
@@ -509,7 +509,7 @@ bool ACesiumSunSky::IsDST(
 
 void ACesiumSunSky::SetSkyAtmosphereGroundRadius(
     USkyAtmosphereComponent* Sky,
-    float Radius) {
+    double Radius) {
   // Only update if there's a significant change to be made
   if (Sky && FMath::Abs(Sky->BottomRadius - Radius) > 0.1) {
     Sky->BottomRadius = Radius;
